added jwt verifications on required routes

+++ src/api/camps/deleteCamp.js
@@ -2,6 +2,9 @@ const Camp = require("../../models/Camps");
 
 const deleteCamp = async (req, res, next) => {
   try {
+    if (req.user.email !== req.query.email) {
+      return res.status(403).send({ message: "Forbidden Access" });
+    }
     const { id } = req.params;
     const camp = await Camp.findByIdAndDelete(id);
     if (!camp) {
