+++ src/api/registeredCamps/deleteRegisteredCamp.js
@@ -2,6 +2,9 @@ const registeredCamps = require("../../models/RegisteredCamps");
 
 const deleteRegisteredCamp = async (req, res, next) => {
   try {
+    if (req.user.email !== req.query.email) {
+      return res.status(403).send({ message: "Forbidden Access" });
+    }
     const { id } = req.params;
     const deletedCamp = await registeredCamps.findByIdAndDelete(id);
     if (!deletedCamp) {
